Add util method handling Spring bean name suffixes
So for automaton we will use a pattern we already used in OpenSAGA a lot
and which is pretty obvious. While the main usage clearly should be
declarative models with added @GraphQLLogic beans, there will be cases
where we want to employ different strategies, offer alternative
implementations for something.

A good way to do that is to create an interface describing the general
contract of the extension point and then have the model JSON contain
references to speaking spring bean names that are defined as spring
beans as part of the local application configuration or predefined by
the AutomatonStandardLogic.

We can then call `applicationContext.getBeansOfType(Class<?> cls)` and we
receive all the spring beans of that type defined in the current spring
configuration mapped by their spring bean name.

It is often useful however to have all Spring beans of the same type
have a common name suffix to both disambiguate them from other spring
beans within the configuration and to make collisions even less likely.

In the model word, where the context of the named bean is clear anyway
however, we don't really want to write the suffixes.

So this commit adds a little helper method that cuts off a given suffix
from all mappings within such a map if the key has such a suffix,
otherwise it is kept as-is.