Use DomainQL 0.20, integrate SchemaDataProvider
So far we've been adding the SchemaDataProvider in the application
WebConfiguration as separate JsViewProvider.

This makes no sense since schema knowledge is a very integral part of
automaton and far from being optional.

So we just create a SchemaDataProvider in AutomatonJsViewProvider and
call it to always provide the schema data (now including generic type
information).

This requires the AutomatonJsViewProvider to know about the current
DomainQL, which generally makes a lot of sense even if we only use it
for the schema data provider now.