AutomatonJsViewProvider: Special case "login" view
"login" is a special view in so far that it is only the login form with
minimal initial data and no process.

For now we just hard-code this. Should we add more exceptional views we
should come up with a better design maybe.

We draw out the process scope provisioning out of "provideCommonData"
into their own method.