InteractiveQuery: Work with untyped fields
This is a problem that showed itself for the DB-View implementation.

When we work with the JOOQ generated Tables, the tables come with a
precisely typed record definition with typed fields.

When we fetch data based on a DSL.table() table however, the records are
not typed and conversion errors may occur because it won't even do
things like convert a long to an int.

Luckily, we have another way to get to this type information because it
is necessarily the property type of the property we're about to set.

So we just change the implementation to use the property type in any
case.