TMP: InteractiveQuery: Reorganize querying
The current code was just plain wrong in some aspects, like following
GraphQL types in the wrong order, only working because our only example
is recursively walking along the same relation.

The default JOOQ join is also not what we really want. It works for our
contrived animals example, but we absolutely need the ability to have
non-null relations, so we change the used join to "LEFT OUTER". We might
make that configurable per relation later.

So with the left outer join, we now potentially have the same problems
we had with disabled groups: fields being null in our results that are
non-null in the schema.

For the disabled thing, without modifying the query, we right-out would
violate non-nullability if we did not substitute.

For the current join scenario it obviously happens because a foreign key
relation is null and a complete object is missing. This is the case we
need to detect, and to do it right, we need to