Implement indirect injection of named queries
With babel-plugin-track-usage@0.3.0 we can use an identifier as argument
for a tracked function and receive that as { __identifier: "MyIdent" }.

We use this feature in automaton to implement named queries.

If a scope contains an definition like

    @observable
    foos = inject( Q_Foo )

the server-side will assume `Q_Foo` to be a properly imported named
query from a module "*/queries/Q_Foo.js" somewhere. It will not verify
that fact.

In Q_Foo.js there must be query() call that defines the actual query and
variables.

This commit detects the map argument we receive and performs a search
 for the "*/queries/Q_Foo" module in the track-usage data and extracts
the actual query parameters.