Refactor WebSocket Handling
We took over some websocket code from earlier prototypes but disabled it
and never really integrated it.

We do so now and whip it into shape.

The old system was using a stupid subclassing approach to messages that
did not work right and generally was stupid.

New system is simple we have two types of messages IncomingMessage and
OutgoingMessage.

    public final class IncomingMessage
    {
        private final String type;
        private final Object payload;
        private final String connectionId;
        private final String messageId;

        ...
    }

    public final class OutgoingMessage
    {
        private final Object payload;

        private final String type;

        ...

    }

The `type` field defines what kind of message it is which client and
server handling their own type namespace.

The `payload` property contains the actual message payload which in the
case of OutgoingMessage can be any svenson compatible POJO. In the case
of IncomingMessage it will be Map/List graph of the incoming JSON data.

The `connectionId` field contains the connection id of the client that
sent the message.

The `messageId` is used to reference the original message if the server
replies to an client message.

Client and server obviously have to agree on what kind of messages
they're dealing with for a request/response message pair, the client
will use `Hub.request({ ... })` to initiate the request and the Hub will
return a promise that is internally associated with the messageId of the
new message.

When the server responds to a request message, it uses a `Response`
payload object.

    public final class Response
    {
        private final String responseTo;

        private final Object reply;

        private final Object error;

        ...
    }

`responseTo` contains the original messageId, `reply` is the response
payload and and `error` is an error that occurred if an error occured
the client will reject the pending promise with an exception built from
the error object.

Enabling Websockets in Automaton
================================

To enable websocket support in automaton, the application needs to set
the property "automaton.enable-websocket" to true and it needs to define
an AutomatonWebsocketHandler bean with the IncomingMessageHandlers to be
used for the application.

Each IncomingMessageHandler handles one type of IncomingMessage. There
is a "GraphQLMessageHandler" that does GraphQL-over-Websocket.