WebsocketHandler: Split connections
There was a persistent problem with Websockets not working after a
login.

A connection was prepared for the login view but never initialized as
the login view just renders a very simple Spring security form without
any other framework participation.

Then the ChatService would broadcast a "user joined" message to
everyone, including the uninitialized session from the login process.

Of course this could just be solved with a safety check, but it seems
preferable to solve the issue in a better way.

This commits splits the connections map into two maps. When a connection
gets registered, it first is put into a `preparedConnections` map and
only moves to `connections` after the connection has been established.
This way it just won't show up in the getConnections() collection unless
it is properly established.

This commit also adds a clean up thread that sleeps for a default 10
minutes then removes all prepared connections that are older than 30
minutes.

Adds a new WebSocketHandlerOptions class and builder to handle these two
options and all further server-side WebSocket configuration options.

Adds a shutdown methods that needs to be called to properly shutdown the
websocket handler.