Ops: Change bean strategy
The @ConditionalOnMissingBean annotation proved too special for our
purposes and apparently requires a full-blown spring-boot extension jar
thing and can't be used in the app.

As a consequence, we need to change the implementation strategy for ops.

This changes the behavior to our previously used pattern of the
automaton standard configuration being incomplete, lacking certain bean
types that have to be defined in every app.

So new rule: The application must define the IdGenerator, StoreOperation
and BatchStoreOperation implementations it needs.

Changes the id generator to be to return a list of ids to enable batch
operation.