Implement multi-phase RuntimeQuery execution
So far we have only supported executing InteractiveQuery based queries
with *-to-one relations, joining together foreign key connected tables.

What was missing was following the foreign keys backwards along a
*-to-many relation.

Using the new relation models this commit refactors the query mechanism
that splits the execution into multiple SQL queries. Whenever a
*-to-many/BackReferenceFetcher based field is navigated, it splits off
a new SQL query that uses IN() or AND/OR constructs based on the previous
SQL queries key fields. The results are then merged together in-memory
and stored inside the already existing FetcherContext to provide the JSON
data in the shape GraphQL expects. So the number of queries is (1 + n)
where n is the number of GraphQL List fields in the query. Each of the
query consist of (1 + m) joined tables, with m being the number of
GraphQL sub-objects entered traversed.

The query mechanism ensures that the necessary key fields are selected
in the query if they are not contained in the user-provided set of
fields.

In contrast to the previous implementation, the fields are selected
unaliased as "tableAlias"."field_name".

Adds two new test-cases for multi-phase querying.