SequenceIdGenerator: Fix and limit
First of all, fetching multiple ids at once wasn't working at all
because we apparently need to alias the fields.

So we create a statement like

SELECT nextval('hs') as id_0, nextval('hs') as id_1, nextval('hs') as id_2

now.

Databases however don't allow unlimited columns. Postgres only allows up
to 1604 columns for example. So we introduce a third constructor
parameter that defines the most columns we will fetch at once.

So if we need 4000 ids and the limit is set to 1600, we will query twice
the max amount of 1600 ids and then the remaining 800.