deleteDomainObject: Add relation cascading
To have a somewhat comfortable way of deleting object graphs based on a
single root object, we add an optional list of relation names to delete
domain object.

If the list of relations is defined, we follow the given relations from
or to the given domain object.

Relation backwards:
-------------------

  Target <-- Other

If another table points to the target object, we delete all entries that
point to our target to be deleted. This is the case that causes a
foreign key constraint violation if it is not done.

Relation forward:
-----------------

  Target --> Other

If we have a one-to-one relation we delete the object the relation is
pointing to. Note that many-to-one relations will still cause an SQL
error if we try to cascade over the many-to-one relation but there are
other objects pointing to the "Other" object.

This direction is rather rare compared to the other one and does not
cause a constraint violation since we're deleting the object with the
offending FK.