Extend FilterDSL for Java-Objects
For interactive queries, we already implemented the FilterDSL module
which provides a fluent API very similar to the JOOQ condition DSL
for the client side.

The fluent API, for example

    and(
        field("a").eq(value("String","Foo")),
        field("b").eq(value("Int",12))
    )

produces a javascript object graph which then can be transferred as JSON
to the server (encapsulated as a Condition scalar within the GraphQL
world).

On the server we then directly translate the map graph to a JOOQ
condition instance which JOOQ then can execute as SQL.

Java-Filter
-----------

Following the way we did things in OpenSAGA, the idea to extend this
Filter API to support other use cases came pretty natural. The
Filter DSL JSON data is an abstract filter definition that can just as
well be applied to Java objects.

The occasion to implement this now is as a filtering mechanism for
pub/sub messages. While this approach may seem oversized for the problem
at hand superficially, the actual needed complexity is not that far
below this. "Send me domain object updates for the types 'Foo' and
'Bar'" is a common pub/sub use case.

And given all the pre-existing java and js code the implementation was
actually pretty easy. Lots of it is auto-generated from the data tables
FilterDSL.js used to construct the JS api at runtime.

The graph of map objects is converted into a filter instance.

    package de.quinscape.automaton.runtime.filter;

    /**
     * Filter node for in-memory filtering on Java objects.
     */
    public interface Filter
    {
        Object evaluate(FilterContext ctx);
    }

There's a filter implementation for all JOOQ conditions and operations
and for value nodes etc.

Lots of again partially generated test cases provide a good coverage of
filter features.