Get rid of OutgoingMessageFactory
The OutgoingMessageFactory and OutgoingMessageFactory method variants of
the websocket handler are remnants of a half-assed inheritance based
version of the outgoing message.

Using simple composition gets rid of the need of that factory so it
really does not serve any purpose now.

Outgoing message payload implementations should just provide a method to
create an outgoing message with that payload and the correct type
property. Either a static method that constructs both payload and
wrapping outgoing message or an instance method that creates an outgoing
message from the current payload instance.