Introduce PubSubService
This commit refactors the both incomplete and under-featured topic
support tacked onto the websocket handler and turns it into a separate
service with additional payload filtering.

After doing some reading on common pub/sub pattern usages, I decided to
go with a mixed topic/content based pub/sub implementation.

In general, the pub/sub messaging keeps being organized by "topics",
which are in the end just a symbolic name describing the purpose of the
server/client live communication.

We now add the possibility for the client to define a filter expression
that is used to filter the messages the client receives. This feature
enables us to cut down on the number of topics and keep the topics very
general, but still being able to control fine-grained what information
out of this general pool we need.

Let's take as an example the upcoming domain monitor which is supposed
to track which objects are currently being edited by other users on a
purely information level for now.

The client is only interested in some of the domain-types for which we
provide monitoring. So with filtering we have just one topic for all
changes of all domain types and the client can just filter out the
domain types relevant to their current process.

PubSubService
-------------

The PubSubService is now a higher-level separate service. The websocket
handler does the general message handling and the pub sub service does
the higher-level association of connections and (filtered) topics.

Alternatively, server-side services can use a TopicListener to subscribe
to topics from the server-side.

There's also a SubscriptionListener to listen to clients subscribing and
unsubscribing.

See de.quinscape.automaton.runtime.topic.PubSubService

Adds two simple pub-sub tests based on a mocked WebSocketSession.