Refactor TopicListener
So for a websocket connection listening to our topic, we need to send an
OutgoingMessage because that's the way our websocket stuff works, but
for the TopicListener, there's no reason to include that outgoing message
wrapper.

The argument is now object. Which is either map-based data or a POJO
depending on whether client-side messages are used. If you publish only
from the server-side, use typed Pojo payloads.

Changes the topic listener interface.

Adjusts some test code.