Implement AlternateStyleProvider
Server-side support for alternate stylesheets.

As it is now, proper HTML standard alternate stylesheets are basically
dead. Some browsers don't offer the user the ability to switch to
alternate styles (or only with plugins) which necessitates script
control, but there is no standard way to deactivate the standard style
and switch to an alternate style.

You can hack something together by setting a disabled attribue on the
link, but MDN advises against that and its not part of the standard.

So we work with only one stylesheet which we switch out with the
alternate stylesheets only by script / a UI-widget.

For the best user experience we need to perform the initial rendering
with the user-defined alternative if there is one.

So we use old-school cookies to store this selection.

This commit adds an AlternateStyleProvider, a JsViewProvider that
defines the available stylesheets and their names.

The provider can be optionally configured with a StyleSheetSelector
that allows a server-side pre-selection of styles (based on current user
or other information).

The provider sets a placeholder variable $CURRENT_STYLE for the
base template.

If there is a selector defined, the style will be the selected style.

Otherwise it will be the first style defined unless we've received a
cookie _AUTO_PREFERRED_STYLE in which case we use that style if the name
is valid.

Example
-------

    @Override
    public void configureViewResolvers(ViewResolverRegistry registry)
    {
        final GraphQLSchema graphQLSchema = domainQL.getGraphQLSchema();
        registry.viewResolver(
            JsViewResolver.newResolver(servletContext, "WEB-INF/template-alternate-styles.html")
                .withResourceLoader(resourceLoader)

                // Process injections and general miscellaneous data we would normally
                // inject by hand in a Spring-JsView application
                .withViewDataProvider(
                    new AutomatonJsViewProvider( ... )
                )

                .withViewDataProvider(
                    new AlternateStyleProvider(
                        servletContext.getContextPath(),
                        Arrays.asList(
                            new StyleSheetDefinition("QS", "/css/bootstrap-automaton.min.css"),
                            new StyleSheetDefinition("QS condensed", "/css/bootstrap-automaton-condensed.min.css")
                        )
                    )
                )
                .build()
        );
    }