Refactor Paging
This is something I should have done sooner but kept pushing because I
thought we might not need it, but we do.

This removes the "currentPage" field from QueryConfig and replaces it
with an "offset" field, the difference besides the name being that the
old field contained the current *page* while the offset field now
contains the actual number of rows skipped, just like SQL does.