Create getDomainTypeIndex end-point
Creates a general-purpose method to create index arrays for the
<Tree.IndexObjects/> component.

The method is given a domain type and a name field within that type and
an optional ConditionScalar.

It returns an array of initial characters present within the (filtered)
domain type result.

Adds a new SimpleFieldResolver implementation that resolves SQL fields
from expressions like "name" or "t2.name". Note that these follow normal
SQL semantics (as opposed to Filter DSL usage within the interactive
query mechanism)