Refactor DefaultPubSubService
Using a numeric id for internally created TopicListener based
registrations seems unnescessary. Registering multiple topic listeners
is no problem in contrast to having to run multiple subscriptions over
one web socket connection.

So we remove the id from the API and use -1 internally.

When we're sending a message to many websocket connections, they're not
exactly the same as they own have their own list of recipient ids, but
the payload is always the same.

We can use that fact to our advantage to wrap the payload inside a
JSONHolder once and then use that JSON holder as message payload to
be reused for every connection.