Implement Domain Monitor as pub sub service
Implement the domain monitor as service using the new pub sub
infrastructure.

The service listens on the "DomainMon" topic and maintains a server-side
log of the domain activity published on that topic. It recasts the JSON
payload to de.quinscape.automaton.model.domainmon.DomainActivity thereby
ensuring a schema on the messages for that topic.

New clients receive the current state matching their registered filter
when they subscribe to the topic.

The monitor makes sure to delete activity when websocket connections
end.

See useDomainMonitor() / useEntity() in automaton-js for the client-side
implementation of this feature.