Enable Scheduling
The use-cases for executing code every once in a while to clean up
things keeps growing.

I already had created a cleanup thread for websocket cleanups.

This commit now enables Spring Boot scheduling and changes the one
cleanup thread to using a @Scheduled annotation.

Minor changes in options as we are now using a fixed check delay of 10
minutes.

We're forced to define an explicit Taskscheduler or the Spring Websocket
internals will substitute a dummy scheduler that throws errors.

We create a ThreadPoolTaskScheduler with a default pool size of 3.

The pool size can be controlled by setting the property

    automaton.scheduler.pool-size