Merge-Service: Solve many-to-many merge
We were so close, but in the end it was one case that broke our back,
which was merging a many-to-many relation OURS which necessitates
updating the link types for which we need the id and the version of the
current "theirs" entries, something that breaks our model of using
"[DomainObject]!" typed GenericScalars due to the limitations of the
DomainObject scalar.

Accepting nested objects would mean to allow DomainObject scalar values
to use output types as input type values. It could be kind of convient
in some situations but it also is *very* odd, and so we avoid it for now
at least.

In the end, it is easier to add extra-information about the link
entities involved.

Adds a new EntityReference type (type, id, version) of which we add an
list to our MergeConflictField.

When requesting the many-to-many tables, we can add the two fields and
add them as entity references.

This allows the client to update their "current" state correctly in all
OURS merge scenarios.