Introduce /_dev/
So far we made an exception for the development GraphQL service so that
it does not need to provide a CSRF token to POST.

Replaces the check for a specific URL with checking if the URI starts
with a /_dev/ prefix. This way we can move all kinds of dev services to
this URI space and disable CSRF protection for all of them.

Renames the request matcher for better fit.