Implement FilterDSL context
So we have a JavaScript FilterDSL that integrates into JOOQ
conditions and can also be used to filter Java objects.

And we are using those filter expressions to control the injection of
interactive query data. This usage has the fundamental limitation that
it can only use static filter expressions, so we already have to express
our filter as raw JSON and cannot use the FilterDSL JavaScript API as of
now :\

Another problem is that these static filter expressions are a bit
limited. Less limited than it might seem at first, because the use case
of "show me the first (filtered) n elements of X" is *very* common.

To increase the flexibility of filter expressions, we add a concept
called filter context.

Filter Context in Java
======================

On the server-side we can define named filter context values which we
can then refer to by-name within the FilterDSL.

The constellation of beans in automaton seemed to like to throw cyclic
dependency errors very easily when configuring this, so in the end I
added an functional interface called FilterContextConfiguration.

The easiest use of it goes like this:

    @Bean
    public FilterContextConfiguration filterContextConfiguration(DSLContext dslContext)
    {
        return registry -> {

            registry.register("param","https://quinscape.de");

            registry.register("userCount",
                ctx -> dslContext.selectCount()
                    .from(APP_USER)
                    .fetchOne()
                    .value1()
            );
        };
    }

Here we see the definition of two filter context values ("param" and
"userCount")

The first example is a simple constant value substitution. This is very
similar if not in practice identical to putting values into the
AppScope, only accessible by filter.

The second is a value factory like method that receives a
FilterContextResolver as context object and that can execute code when
the value is needed.

Caching
=======

The server implementation makes sure that these factory methods are not
invoked excessively. The API allows for the definition of a value
lifetime for each user of the filter context API.

Filters converted to JOOQ conditions will be invoked at most once per condition
transform.

The PubSubService will evaluale each filter context value at most once
per published message.

This applies to a context value being present in a filter expression
more than once as well as values that are dependent on other values.

The tests define such a construct for example.

        registry.register("auth", ctx -> {
            return AutomatonAuthentication.current();
        });
        registry.register("userId", ctx -> {

            AutomatonAuthentication auth = ctx.resolveContext("auth");
            return auth.getId();
        });
        registry.register("login", ctx -> {

            AutomatonAuthentication auth = ctx.resolveContext("auth");
            return auth.getLogin();
        });

Both userId and login depend on another value "auth". Now in this
example it wouldn't matter if we'd execute these factory methods over
and over but there might be actual costly queries behind those values in
practice.

Filter Context in JavaScript
============================

On the FilterDSL side, we can reference the filter context values by
name.

    import { FilterDSL } from "@quinscape/automaton-js"

    // deconstruct FilterDSL methods
    const { field, context } = FilterDSL;

    or(
        field("url")
            .eq(
                context("param")
            ),
        field("count")
            .lt(
                context("userCount")
            )
    )

Programmatic Access
===================

If you should need the current filter context values, access the
registered FilterContextRegistry bean. Consider wrapping multiple
consecutive uses with a CachedFilterContextResolver.

If you should need filter context values in JavaScript outside of a
filter, you can access the resolveFilterContext GraphQL method

    type QueryType
    {
        resolveFilterContext(names: [String]!): [GenericScalar]
    }

which will give you a list of generic scalar instances for a list of
filter context value names.