Simplify AutomatonJsViewProvider config for app
The process of constructing the AutomatonJsViewProvider within the app's
WebConfiguration was already complicated and gets more and more
complicated.

We streamline this a little in two ways:

First we move the construction of the actual AutomatonJsViewProvider
into the lib and then just inject it.

Seeing how we increasingly had references to other services just for
them to be regarded during initial data block generation, we use a more
simple approach where we just pick up all AutomatonJsViewProvider
defined as Spring beans within automaton and call all of them within
AutomatonJsViewProvider.

Note that this only seems to work within the library. If you have your
own additional js view providers in your app, you should stick to
WebConfiguration to define them.