Implement UserInfoService
So far, we only had the core properties of the authentication mechanism
available as information on the user. This included the user database
id, but still necessitated querying the information when needed.

We now introduce the UserInfoService which fetches and caches additional
informational data on a user after they have logged in successfully.

The UserInfoService works with a UserInfoProvider interface that allows
arbitrary customization for the user info data fetching.

As default implementation we provide a class IQueryUserInfoProvider
that executes a filtered interactive query based GraphQL query.

We also provide an abstract base class to easily implement alternate
GraphQL-based user info retrieval.

We define a JsViewProvider bean that automatically inserts a "userInfo"
block into the intial data when a UserInfoService is defined.

ExampleUsage
------------

Within the ServiceConfiguration of the app:

    @Bean
    public UserInfoService userInfoService(
        @Lazy DomainQL domainQL
    )
    {
        return new UserInfoService(
            new IQueryUserInfoProvider(
                domainQL,
                // language=GraphQL
                "query getUserInfo($config: QueryConfigInput!)\n" +
                "{\n" +
                "    iQueryAppUser(config: $config)\n" +
                "    {\n" +
                "        rows{\n" +
                "            id\n" +
                "            created\n" +
                "            lastLogin\n" +
                "            foos{\n" +
                "                name\n" +
                "            }\n" +
                "            bazes{\n" +
                "                name\n" +
                "            }\n" +
                "        }\n" +
                "    }\n" +
                "}"
            )
        );
    }