Add automaton types
Use the new domainql 2.6 --merge feature to merge the current library
javadocs with handwritten docs about future types.

We basically add type documentation for all the base database types
automaton expects within the application.

Even though the concrete types are yet known including the final java
packages and fields and relations, we know the names of the parts that
are standard and can document that.

In the end they will end up overwriting the standard "Generated from
public.xyz" documentation for those types in the applications.