Implement attachments
Implement a basic system for file attachments. We split up the
data-storage into two parts: one for core attachment data and one for
the payload data.

AttachmentRepository
====================

The attachment repository provides access to the core attachment data
which is always stored in the database. We expect a table matching

    type AppAttachment {
        "Description of the attachment, might be former filename"
        description: String
        "attachment id"
        id: String!
        "Attachment media type"
        type: String!
        "Attachment url as alternative to binary content"
        url: String
    }

public interface AttachmentRepository
{
    Attachment getAttachment(String attachmentId);
    void store(Attachment attachment, byte[] data);
    void delete(String attachmentId);

    AttachmentContentRepository getContentRepository();
}

We provide a default implementation for attachment repository.

AttachmentContentRepository
===========================

The default attachment repository is created with an attachment content
repository that handles the storage of the actual attachment data.

There are two implementations of the it.

DatabaseAttachmentContentRepository
-----------------------------------

The DatabaseAttachmentContentRepository stores the attachment contents
in another database table with a foreign key pointing to app_attachment.

    "Data-base backed attachments"
    type AppAttachmentData {
        "DB column 'attachment_id'"
        attachmentId: String!
        "attachement as binary data"
        data: [Byte]!
        "attachment data id"
        id: String!
    }

Example configuration:

    @Bean
    public AttachmentRepository attachmentRepository(
        DSLContext dslContext,
        @Lazy DomainQL domainQL
    )
    {
        return new DefaultAttachmentRepository(
            dslContext,
            domainQL,
            new DatabaseAttachmentContentRepository(
                dslContext,
                domainQL
            )
        );
    }

FileBasedAttachmentContentRepository
====================================

Alternative we have the FileBasedAttachmentContentRepository which
stores the attachment data inside a file directory.

Example configuration:

    @Bean
    public AttachmentRepository attachmentRepository(
        DSLContext dslContext,
        @Lazy DomainQL domainQL,
        @Value("${myapp.attachments}") String attachmentDirectory
    )
    {
        return new DefaultAttachmentRepository(
            dslContext,
            domainQL,
            new FileBasedAttachmentContentRepository(
                new File(attachmentDirectory)
            )
        );
    }

Here is an example using the Java property myapp.attachments to define
the location of the file attachment storage directory.