Refactor UserInfo
The userInfo mechanism still had one serious defect, which is that is
wasn't converted on the client side like it should be.

But to convert it, we need to know the type.

So we refactor the IQueryUserInfoProvider so that it returns an instance
of a new internal class UserInfo which contains the type and the GraphQL
provided userInfo blob as "info".

For this purpose it is now important that the GraphQL query queries the
"type" field of the InteractiveQueryXXX object so that the client side
knows how to convert.

(We only have the _type fields *after* conversion)