Refactor Filter transformers
The condition builder refactoring on the client side has some minor
implications for the server side.

We need a "false" condition node now to deal with conditional not()s. So
we add both "true" and "false" in both JOOQ transformer and Java
transformer

Fixes some minor problems with null handling and missing operands
arrays.