MergeServiceImpl: Fix loading behavior
The lazy loading of DB-versions needs to work for
findVersionByPrevious, too.

We first try to find in-memory. The only way for an in-memory entry to
disappear is currently the global flush, so if the previous version is
there, all versions before it will be there.

If we don't find it, then we try loading it. This should happen only on
service restart or after a global flush. It also happens in the tests.