MergeServiceImpl: Fix insertion order
We need to make sure that our inserts happen in the right order.

This commit adds a resolveChangeDependencies methods that analyzes the
changes.

For every change of a foreign key field, it makes sure that changes to
the same entity happen before the foreign key is changed.

The integration test in automaton-test tests with with a main entity and
two connected new related instance.