Ignore non many-to-many backwards relations
If we go through the type for merge info, don't assume that all List
fields are of many-to-many relations as they may just as well be other
cases of a backwards followed fk resulting in a list of objects.

So instead of throwing an exception, we quietly ignore those relations
for many-to-many purposes and only record the relations for which we
indeed have found the two outgoing relations of the many-to-many link
type.