Formalize simple views
So far we only had one case where we used a simple view without any of
the process logic and that is the login view. The provisions was
hard-coded with an exception for the login bundle.

This commit generalizes the concept. We define that every entry-point
whose name starts with "v-" is a simple view for which the process
preparation will be skipped.

We provide a general @RequestMapping "/v/{viewName/**" to address the
simple views.

Application changes
===================

Applications no longer need to define the "login" bundle mapping.

The .loginPage() call in the SecurityConfiguration has to be changed to
"/v/v-login/" to fit the new mapping.

Correspondingly, the login entry point defined in the webpack.config.js
must be called "v-login" now.