Refactor for new DomainQL meta data
Contains the necessary changes in automaton to work with the new
DomainQL meta data.

We remove two js view providers and convert them into one
AutomatonMetadataProvider.

The former config.decimalPrecision is now a field meta
"decimalPrecision" on that field.

For example

    inputSchema.getFieldMeta("Garply", "opt", "decimalPrecision")
    ( see automaton-test )

The fieldLengths block is now a "maxLength" field meta.

inputSchema.getFieldMeta("Foo", "name", "maxLength")

For database originated types, the maxLength still automatically
reflects the maximum length on the database column. If you need it for
handwritten Java-classes you can use the same @Column annotation

Note that the nameFields already are a type meta now in DomainQL.

The DecimalPrecision POJO is simplified to only contain precision and
scale properties. we no longer need the FieldLength POJO.