IQuery: Tolerate computed fields in queries
The new @GraphQLComputed fields have no equivalent columns in the
database but instead mash up several other columns in memory.

We make sure that we can handle those missing columns to execute such a
query and have the GraphQL fetchers use the @GraphQLComputed methods
to produce and query the computed fields from the other SQL columns.

We cannot sort or filter these computed columns. We could have
implemented soem half-assed bad way of doing that but decided against it
in the end because ther user reality coming from that was just too plain
horrible.

To sort or filter based on a computed column the query has to
reimplement the @GraphQLComputed method as FilterDSL condition.

For example:

    <DataGrid.Column
        name="extra"
        filter={
            val => field("name")
                .concat(value(':'))
                .concat(field('num'))
                .containsIgnoreCase(value(val, 'String'))
        }
        sort={ field("name")
            .concat(value(':'))
            .concat(field('num'))  }
    />