Generalize custom JOOQ transformers
We already had one case where our FilterDSL expressions diverge from
JOOQ and that is "toString()" which is a cast to String.class, which we
can't really express in the FilterDSL graph language yet. Same reason we
did continue with "like" in AUTO-37.

This refactor at least goes half the way and introduces a way to enhance
FilterDSL functionality beyond what we can do with JOOQ's Field methods.

For now that is only toString() and concat().

The chosen implementation of passing in the recursive transform function
enables local FilterDSL rules which might enable more straight-forward
local solutions to the filter transformation argument problem mentioned
above, compared to a generalized solution.