FilterDSL: Increase security by positive list
We are indirectly using reflection for to translate the FilterDSL name
fields of  "Condition" and "Operation" typed nodes into JOOQ
conditions.

It is not totally impossible that there are names that would have
security implications now or in future JOOQ versions.

To nail this down and in the spirit of being extra careful with
user-input, we introduce a positive list of the same functionality we
support in the FilterDSL.js module in @quinscape/automaton-js.